refactor: add 'sort-import' rule for eslint
