+++ src/components/Converter/Output/PromptMessage.tsx
@@ -1,7 +1,6 @@
+import CardFormatContext from '@contexts/CardFormat';
 import { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
-
-import { CardFormatContext } from './context';
 
 const PromptMessage = ({
   hasConvertedData,
