+++ src/contexts/Toast.ts
@@ -1,11 +1,7 @@
+import { ToastContextProps } from '@interfaces/App';
 import { createContext } from 'react';
 import { toast } from 'react-toastify';
 
-import { NotifyType } from '@customTypes/App';
-
-interface ToastContextProps {
-  notify: (msg: string, type?: NotifyType) => void;
-}
 
 const ToastContext = createContext<ToastContextProps>({
   notify: (msg: string) => {
@@ -13,4 +9,4 @@ const ToastContext = createContext<ToastContextProps>({
   },
 });
 
-export { ToastContext };
+export default ToastContext;
