+++ src/interfaces/App.ts
@@ -1,0 +1,7 @@
+import { NotifyType } from '@customTypes/App';
+
+interface ToastContextProps {
+  notify: (msg: string, type?: NotifyType) => void;
+}
+
+export type { ToastContextProps };
