+++ src/shared/UI/Loader/index.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
 import './style.css';
+import React from 'react';
 
 const Loader: React.FC = () => {
   return (
