+++ src/tests/components/Converter/Input/index.test.tsx
@@ -1,8 +1,8 @@
-import { render, fireEvent } from '@testing-library/react';
-import { describe, it, expect, vi } from 'vitest';
-import ConverterInput from '../../../../components/Converter/Input';
+import { describe, expect, it, vi } from 'vitest';
+import { fireEvent, render } from '@testing-library/react';
+import ConverterInput from '@components/Converter/Input';
 import { I18nextProvider } from 'react-i18next';
-import i18n from '../../../../i18n';
+import i18n from '@shared/i18n';
 
 const mockProps = {
   labelName: 'test-label',
