+++ src/tests/components/LanguageToggler.test.tsx
@@ -1,10 +1,10 @@
-import { render, screen, fireEvent } from '@testing-library/react';
-import { describe, it, expect } from 'vitest';
-import { Provider } from 'react-redux';
-import LanguageToggler from '../../components/LanguageToggler/index';
-import store from '@store/index';
+import { describe, expect, it } from 'vitest';
+import { fireEvent, render, screen } from '@testing-library/react';
 import { I18nextProvider } from 'react-i18next';
-import i18n from '../../i18n';
+import LanguageToggler from '../../components/LanguageToggler/index';
+import { Provider } from 'react-redux';
+import i18n from '@shared/i18n';
+import store from '@store/index';
 
 describe('LanguageToggler Component', () => {
   it('should render the LanguageToggler component', () => {
