+++ vite.config.ts
@@ -1,16 +1,17 @@
 /// <reference types="vitest/config" />
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
 import path from 'path';
+import react from '@vitejs/plugin-react';
 
 export default defineConfig({
   plugins: [react()],
   base: './',
   resolve: {
     alias: {
-      '@src': path.resolve(__dirname, 'src'),
       '@components': path.resolve(__dirname, 'src/components'),
+      '@contexts': path.resolve(__dirname, 'src/contexts'),
       '@shared': path.resolve(__dirname, 'src/shared'),
+      '@tools': path.resolve(__dirname, 'src/tools'),
       '@store': path.resolve(__dirname, 'src/store'),
       '@interfaces': path.resolve(__dirname, 'src/interfaces'),
       '@customTypes': path.resolve(__dirname, 'src/types'),
